bigtable/requirements.txt
google-cloud-bigtable==2.3.1
google-cloud-core==1.7.2